Moved reusable handler classes to their own directory
