+++ src/IntegrationHandler.php
@@ -3,6 +3,8 @@
 namespace Inpsyde\MultilingualPress2to3;
 
 use Dhii\I18n\StringTranslatingTrait;
+use Inpsyde\MultilingualPress2to3\Handler\HandlerInterface;
+use Inpsyde\MultilingualPress2to3\Handler\HandlerTrait;
 use Psr\Container\ContainerInterface;
 
 /**
